pyproject.toml
[project]
name = "emotion-machine-interview"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]

requires-python = ">=3.10,<3.12"

dependencies = [
    "annotated-types>=0.7.0,<0.8.0",
    "certifi>=2024.2.0,<2025.0.0",
    "charset-normalizer>=3.3.2,<4.0.0",
    "filelock>=3.13.1,<4.0.0",
    "fsspec>=2024.6.1,<2025.0.0",
    "idna>=3.6,<4.0",
    "jinja2>=3.1.6,<4.0.0",
    "joblib>=1.3.2,<2.0.0",
    "markupsafe>=3.0.3,<4.0.0",
    "packaging>=24.0,<25.0",
    "pillow>=10.2.0,<12.0.0",

    "pydantic>=2.12.5,<3.0.0",
    "pydantic-core>=2.41.5,<3.0.0",

    "pyyaml>=6.0.1,<7.0.0",
    "regex>=2024.4.16,<2025.0.0",
    "requests>=2.32.3,<3.0.0",
    "setuptools>=69.5.1,<70.0.0",
    "threadpoolctl>=3.2.0,<4.0.0",
    "typing-inspection>=0.4.2,<0.5.0",
    "typing-extensions>=4.12.2,<5.0.0",
    "urllib3>=2.2.2,<3.0.0",

    "numpy>=2.0.0,<3.0.0",
    "pandas>=2.2.1,<3.0.0",
    "matplotlib>=3.8.3,<4.0.0",
    "seaborn>=0.13.2,<0.14.0",
    "scipy>=1.12.0,<2.0.0",
    "scikit-learn>=1.4.2,<2.0.0",

    "notebook>=7.1.1,<8.0.0",

    "hf-xet>=1.2.0,<2.0.0",
    "huggingface-hub>=0.25.2,<1.0.0",
    "safetensors>=0.4.3,<1.0.0",

    "tokenizers>=0.19.1,<0.20.0",
    "transformers>=4.43.4,<5.0.0",
    "sentence-transformers>=2.7.0,<3.0.0",

    "tqdm>=4.66.4,<5.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "mpmath>=1.3.0,<2.0.0",
    "openai (>=2.9.0,<3.0.0)",
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "faiss-cpu (>=1.13.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
